Refactored Shooters to separate NetworkPrefab & LocalPrefab
