+++ TP2_Engin2/Assets/Prefabs/ShooterLocal.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 252eb8d7e9c40b14b82145b3b3c37089
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
