+++ TP2_Engin2/Assets/Prefabs/ShooterNetwork.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cb6272c282e142149892cc6f901ecf0c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
