+++ TP2_Engin2/Assets/Scripts/LevelCharacter/LocalLevelPlayerController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec8c163b3b609834e96954c866f89c0e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
