+++ TP2_Engin2/Assets/Scripts/MainCharacterScripts/CharacterSM/RunnerSM.cs
@@ -243,7 +243,7 @@ public class RunnerSM : GM_BaseStateMachine<CharacterState>
         Animator.SetBool("IsGrounded", IsInContactWithFloor());
     }
 
-    public void SetCharacterGo(GameObject go)
+    public void SetParentGo(GameObject go)
     {
         MC = go;
     }
